[api] make distribution its own resource
