+++ datadog/threadstats/reporters.py
@@ -15,7 +15,7 @@ class Reporter(object):
 class HttpReporter(Reporter):
 
     def flush_distributions(self, distributions):
-        api.Metric.send_dist(distributions)
+        api.Distribution.send(distributions)
 
     def flush_metrics(self, metrics):
         api.Metric.send(metrics)
